[dbt] Preserve the WARNING file at sync

+++ cmd/sync.go
@@ -194,7 +194,7 @@ func runSync(cmd *cobra.Command, args []string) {
 	if content != nil {
 		for _, info := range content {
 			fullPath := path.Join(depsDir, info.Name())
-			if !done[fullPath] && fullPath != workspaceModuleSymlink {
+			if !done[fullPath] && fullPath != workspaceModuleSymlink && info.Name() != util.WarningFileName {
 				log.Log("Deleting '%s'\n", fullPath)
 				os.RemoveAll(fullPath)
 			}
